Improves: rols functionality included, icons for actions added

+++ app/Http/Controllers/ProfileController.php
@@ -13,15 +13,21 @@ use Inertia\Response;
 
 class ProfileController extends Controller
 {
+    protected $admin = 1;
     /**
      * Display the user's profile form.
      */
     public function edit(Request $request): Response
     {
-        return Inertia::render('Profile/Edit', [
-            'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
-            'status' => session('status'),
-        ]);
+        $user_rol = Auth::user()->rol;
+        if ($user_rol == $this->admin) {
+            return Inertia::render('Profile/Edit', [
+                'mustVerifyEmail' => $request->user() instanceof MustVerifyEmail,
+                'status' => session('status'),
+            ]);
+        } else {
+            return Inertia::render('NoPermissions');
+        }
     }
 
     /**
@@ -29,14 +35,16 @@ class ProfileController extends Controller
      */
     public function update(ProfileUpdateRequest $request): RedirectResponse
     {
-        $request->user()->fill($request->validated());
+        $user_rol = Auth::user()->rol;
+        if ($user_rol == $this->admin) {
+            $request->user()->fill($request->validated());
 
-        if ($request->user()->isDirty('email')) {
-            $request->user()->email_verified_at = null;
+            if ($request->user()->isDirty('email')) {
+                $request->user()->email_verified_at = null;
+            }
+
+            $request->user()->save();
         }
-
-        $request->user()->save();
-
         return Redirect::route('profile.edit');
     }
 
@@ -49,15 +57,17 @@ class ProfileController extends Controller
             'password' => ['required', 'current_password'],
         ]);
 
-        $user = $request->user();
+        $user_rol = Auth::user()->rol;
+        if ($user_rol == $this->admin) {
+            $user = $request->user();
 
-        Auth::logout();
+            Auth::logout();
 
-        $user->delete();
+            $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
-
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+        }
         return Redirect::to('/');
     }
 }
